refactor: main function so we can use it in tests
